CRUD with Validations completed

+++ src/main/java/com/hardik/customer/CustomerService.java
@@ -1,6 +1,7 @@
 package com.hardik.customer;
 
 import com.hardik.exception.DuplicateResourceException;
+import com.hardik.exception.RequestValidationException;
 import com.hardik.exception.ResourceNotFound;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
@@ -44,6 +45,36 @@ public class CustomerService {
     }
 
     public void deleteCustomer(Integer customerId){
+        if(!customerDao.existsPersonWithId(customerId)){
+            throw new ResourceNotFound("Customer with Id [%s] not found.".formatted(customerId));
+        }
         customerDao.deleteCustomer(customerId);
+    }
+
+    public void updateCustomer(Integer customerId, CustomerUpdateRequest updateRequest){
+        Customer customer = getCustomer(customerId);
+        boolean changes = false;
+        if(updateRequest.name()!=null && !customer.getName().equals(updateRequest.name())){
+            customer.setName(updateRequest.name());
+            changes = true;
+        }
+
+        if(updateRequest.email()!=null && !customer.getEmail().equals(updateRequest.email())){
+            if(customerDao.existsPersonWithEmail(updateRequest.email())){
+                throw new DuplicateResourceException("Email already in use.");
+            }
+            customer.setEmail(updateRequest.email());
+            changes=true;
+        }
+
+        if(updateRequest.age()!=null && !customer.getAge().equals(updateRequest.age())){
+            customer.setAge(Integer.valueOf(updateRequest.age()));
+            changes=true;
+        }
+
+        if(!changes){
+            throw new RequestValidationException("No data change captured.");
+        }
+        customerDao.updateCustomer(customer);
     }
 }
